requirements.txt
absl-py==0.9.0
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.4
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
astropy==4.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
Automat==20.2.0
Babel==2.8.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.1
cchardet==2.1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.10.1
decorator==4.4.1
defusedxml==0.6.0
demoji==0.1.5
distributed==2.10.0
docutils==0.16
elasticsearch==7.5.1
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastcache==1.1.0
Flask==1.1.1
fsspec==0.6.2
future==0.18.2
gast==0.2.2
geographiclib==1.50
geopy==1.21.0
gevent==1.4.0
gmpy2==2.0.8
google-pasta==0.1.8
googletransx==2.4.2
GPUtil==1.4.0
greenlet==0.4.15
grpcio==1.16.1
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hyperlink==17.3.1
hypothesis==5.5.4
idna==2.8
idna-ssl==1.1.0
imageio==2.6.1
imagesize==1.2.0
imbalanced-learn==0.6.2
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rl2==1.0.3
keyring==21.1.0
kiwisolver==1.1.0
langdetect==1.0.8
lazy-object-proxy==1.4.3
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.7.5
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.2.3
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.6.1
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Protego==0.1.16
protobuf==3.11.4
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
py-googletrans==1.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycares==3.1.1
pycld2==0.41
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
PyDispatcher==2.0.5
pyflakes==2.1.1
Pygments==2.5.2
PyHamcrest==2.0.2
pylint==2.4.4
pymongo==3.10.1
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.16.0
ruamel-yaml==0.15.87
schedule==0.6.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
Scrapy==2.0.0
seaborn==0.10.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.13
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tb-nightly==1.14.0a20190603
tblib==1.6.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
tf-estimator-nightly==1.14.0.dev2019060501
toolz==0.10.0
tornado==6.0.3
tqdm==4.45.0
traitlets==4.3.3
tweepy==3.8.0
-e git+https://github.com/twintproject/twint.git@8cec66a256940d3daf9a9d34a5da8a61f26fe165#egg=twint
Twisted==19.10.0
typed-ast==1.4.1
typing==3.7.4.1
typing-extensions==3.7.4.1
unicodecsv==0.14.1
urllib3==1.25.8
w3lib==1.20.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.11.2
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwings==0.17.1
xlwt==1.2.0
yarl==1.4.2
zict==1.0.0
zipp==2.2.0
zope.interface==4.7.1
